refactor: extract TaskDetailEditForm from task-details-sheet (#24)
Move all edit state and edit-mode JSX into TaskDetailEditForm.
Reuses TaskSubtasksEditor and TaskTagsInput from task-form/.
Reduces task-details-sheet from 597 to ~175 lines.
